refactor: `VerifiWASM` and `WasmVerify` monads fail silently
Previously, errors within the two monads were duplicated, since they are written when running the monads and then a Haskell error was being thrown. With these changes, now we are not throwing a Haskell error, we are only exitting with exit code 1 (the logs from the monads are still kept and printed).